Disable adding publisher to XML for DataCite
In #16 (specifically cbf2a2cbf042735089ae817133ebf6c6bb8e8100) the codebase was
changed to forcibly add the "Default publisher" value to all submissions as it
wasn't being (reliably) added. In the rework for 2.1.0 (I'm looking at
fd41e34efc7cc00a529c8f45ad22bb2b376d7768 but its other commits are relevant
too) the behaviour changed and now publisher is added if available.

This has resulted in the unfortunate situation that:
* If you have no publisher, none is added
* If you do have a publisher, a second is added

Not the situation we want:
* If you have no publisher, one is added
* If you have a publisher, nothing changes