Change datacite_mapping_date and date validation
Firstly (in terms of execution order) is a change to the date validation to
make explicit the intent - that date and date type must be set and date type
must equal published.
I believe this also resolves a matching bug.

Secondly is a relaxing of the date requirements fro datacite_mapping_date to
stop checking date_type is published. If date type is was unset this would pass
validation then cause an error to be printed due to an empty value being used
in string equality testing.
Further tightening the check in datacite_mapping_date was a possibility but
since validation is run first it seems reasonable that datacite_mapping_date be
more trusting (and may even make things more flexibile for those who customise
their validation).